machine_learning/requirements.txt
langchain==0.1.12 # for LLM integration
openai==1.12.0 # for OpenAI API
pinecone-client==3.0.2 # for vector database
sentence-transformers==2.5.1 # for sentence embedding (which is used to convert text into vectors)
python-dotenv==1.0.1 # for environment variables
pandas==2.2.1 # for data manipulation
numpy==1.26.4 # for numerical operations
psycopg2-binary==2.9.9  # for PostgreSQL connection
tiktoken==0.6.0  # for token counting with OpenAI 